Advertiser test case (#184)
The test case was broken by the interface changes
to the GenericMessage. The endpoint access via a
property was causing the object to be returned,
not its value.

The test was not failing, but it could not get
any data.

The inventory class has also been moved out of
the decoder class.