Advertiser message: interface compliance (#200)
The adv message was not complying with the generic
message interface. The contents of the APDU were left
out of the message.apdu dict, thus the serialization
was only providing the basic packet information.

This change also updates the test case to make use 
of the message serialization method. A dashboard
to help with the test case results is also added to 
the examples/dashboards folder.