HTTP interface (#224)
This updates adds cache control to the HTTP API
and refactors the code to ease up changing the HTTP
verb. A simple interface test is present in the tests folder.

Logging has been rethought in order to ease up ingestion
through fluentd.  A simple kibana dashboard is available 
from the examples folder.