Update gw cli (#236)
Adds a common method to define and acquire the command options.
The options are set through a dictionary, where one can define
the type and default value.

The command line has been refactored into a separate module. 
The goal of this split is to make it easier to manage
the gateway specific commands as well as to allow other cli
to be built more easily. 

The gateway commands live in cli/gateway, while the
common shell is living in cli/shell.

Closes #102
Closes #228
Closes #167
Closes #239
Closes #242